Move yq to it's own package
