+++ cli/pkg/yq/yq_test.go
@@ -1,4 +1,4 @@
-package utils
+package yq
 
 import (
 	"os"
